Shared parameter documentation

Shared parameter documentation

Arguments

object
Object.

x
Object.

y
Object.

BPPARAM
bpparamClass. BiocParallel parameter to specify the desired processor configuration.
We recommend using one of the following:

bpparam.

SerialParam.

MulticoreParam.

alpha
numeric(1) or NULL. Adjusted P value ("alpha") cutoff. If left NULL, will use the cutoff defined in the object.

asis
logical(1). Set this to TRUE when using the function inside a loop or inside an R Markdown chunk with 'results = "asis"' enabled.

aspectRatio
integer(1). Aspect ratio.

assay
vector(1). Assay name or index position.

assays
SimpleList. Count matrices, which must have matching dimensions. Counts can be passed in as either a dense matrix (matrix) or sparse matrix (sparseMatrix).

base
integer(1). Logarithm base.

borderColor
character(1) or NULL. Border color.

captions
character or NULL. Captions. If NULL, the names of the primary input list will be used automatically.

cells
character. Cell identifiers.

censorSamples
character. Specify a subset of samples to censor.

class
character(1). Object class.

classes
character. Object classes.

closed
logical(2). Should the lower (1) and upper (2) bounaries be closed?

colData
DataFrame. Metadata describing the assay columns. For bulk RNA-seq, this data describes the samples. For single-cell RNA-seq, this data describes the cells.

colnames
logical(1). Apply to column names.

color
ScaleDiscrete. Desired ggplot2 color scale. Must supply discrete values. When set NULL, the default ggplot2 color palette will be used. If manual color definitions are desired, we recommend using ggplot2::scale_color_manual().

To set the discrete color palette globally, use:

options(acid.color.discrete = ggplot2::scale_color_viridis_d())

compress
logical(1) or character(1). These character strings are currently allowed for save(): "gzip", "bzip2", or "xz".

counts
matrix. Count matrix. Normalized counts are recommended.

countsAxisLabel
character(1). Counts axis label.

dark
logical(1). Plot against a dark background using the acid_theme_light() ggplot2 theme.

dims
integer. Vector of length 2 that denotes the columns from the reduced dimension matrix to use for centerX and centerY column calculations. Defaults the first and second dimensions.

dir
character(1). Directory path.

direction
character(1). Plot "both", "up", or "down" directions.

ensemblRelease, release
integer(1). Ensembl release version (e.g. 90). We recommend setting this value if possible, for improved reproducibility. When left unset, the latest release available via AnnotationHub/ensembldb is used. Note that the latest version available can vary, depending on the versions of AnnotationHub and ensembldb in use.

envir
environment. Environment.

expression
character(1). Calculation to apply. Uses match.arg() internally and defaults to the first argument in the character vector.

FUN
function.

file
character(1). File path.

fill
ggproto/ScaleDiscrete. Desired ggplot2 fill scale. Must supply discrete values. When set to NULL, the default ggplot2 color palette will be used. If manual color definitions are desired, we recommend using ggplot2::scale_fill_manual().

To set the discrete fill palette globally, use:

options(acid.fill.discrete = ggplot2::scale_fill_viridis_d())

flip
logical(1). Flip x and y axes. Recommended for plots containing many samples.

fun
function.

gene2symbol
Gene2Symbol. Gene-to-symbol mappings. Must contain geneID and geneName columns. See Gene2Symbol for more information.

genes
character. Gene identifiers.

genomeBuild
character(1). Ensembl genome build assembly name (e.g. "GRCh38"). If set NULL, defaults to the most recent build available. Note: don't pass in UCSC build IDs (e.g. "hg38").

gffFile
character(1). GFF/GTF (General Feature Format) file. Generally, we recommend using GTF (GFFv2) instead of GFFv3.

geom
character(1). Plot type. Uses match.arg() internally and defaults to the first argument in the character vector.

headerLevel
integer(1) (1-7). Markdown header level.

i
Indices specifying elements to extract or replace. Indices are numeric or character vectors, empty (missing), or NULL.

For more information:

help(topic = "Extract", package = "base")

ignoreTxVersion
logical(1). Don't the include the transcript version in the identifier. Only applies when level = "transcripts". This simplifies identifier matching when generating a tx2gene file.

inherits
logical(1). Should the enclosing frames of the environment be searched?

interestingGroups
character. Groups of interest to use for visualization. Corresponds to factors describing the columns of the object.

j
Indices specifying elements to extract or replace. Indices are numeric or character vectors, empty (missing), or NULL.

For more information:

help(topic = "Extract", package = "base")

label
logical(1). Superimpose sample text labels on the plot.

labelSize
integer(1). Size of the text label.

labels
list. ggplot2 labels. See ggplot2::labs() for details.

lfcThreshold
numeric(1) or NULL. Log fold change ratio (base 2) cutoff threshold. If left NULL, will use the cutoff defined in the object.

level
character(1). Return as genes or transcripts.

legend
logical(1). Include plot legend.

limit
numeric(1). Threshold limit.

lower
numeric(1). Lower boundary.

max
numeric(1). Recommended maximum value cutoff.

metadata
list. Metadata.

metric
character(1). Metric type, either length or elements.

min
numeric(1). Recommended minimum value cutoff.

minCounts
integer(1). Minimum number of counts per feature (i.e. gene).

msg
NULL or character(1). Custom message to return.

n
integer(1). Number to include.

ntop
integer(1). Number of top features to label.

nullOK
logical(1). If set to TRUE, x may also be NULL.

ordered
logical(1).

organism
character(1). Full Latin organism name (e.g. "Homo sapiens").

overwrite
logical(1). Overwrite existing file on disk.

pattern
character(1). Pattern to use for matching.

perMillion
logical(1). Display as counts per million.

perSample
logical(1). Visualize the distributions per sample.

pointAlpha
numeric(1) (0-1). Alpha transparency level.

pointColor
character(1). Default point color for the plot.

pointScalar
integer(1). Default point size for the plot.

pointSize
numeric(1). Point size for dots in the plot. In the range of 1-3 is generally recommended.

pointsAsNumbers
logical(1). Plot the points as numbers (TRUE) or dots (FALSE).

prefilter
logical(1). Apply prefiltering. Remove zero count genes.

progress
logical(1). Show progress, using progress bars.

projectDir
character(1). Project directory path.

reducedDims
SimpleList. List containing matrices of cell coordinates in reduced space.

reduction
vector(1). Dimension reduction name or index position.

removeNA
logical(1). Remove NA values from calculations.

return
character(1). Return type. Uses match.arg() internally and defaults to the first argument in the character vector.

rowData
DataFrame. Metadata describing the assay rows, if genomic ranges are not available. Use rowRanges (GRanges) instead, if possible.

rowRanges
GRanges or GRangesList. Genomic ranges (e.g. genome annotations). Metadata describing the assay rows.

rownames
logical(1). Apply to row names.

sampleMetadataFile
character(1). Sample metadata file path. CSV or TSV is preferred, but Excel worksheets are also supported. Check the documentation for conventions and required columns.

samples
character. Sample identifiers.

sigPointColor
character. Color names for labeling upregulated and downregulated genes. Also supports a character string for labeling DEGs with the same color, regardless of direction.

sort
logical(1). Resort using sort.

spikeNames
character. Vector indicating which assay rows denote spike-in sequences (e.g. ERCCs).

subtitle
character(1). Subtitle.

tabset
logical(1). Include tabset marker.

text
character(1). Text.

title
character(1). Title.

traceback
logical(1). Include traceback in error message. See traceback() for details. rlang::entrace() also works nicely and can be set in .Rprofile.

trans
character(1). Name of the axis scale transformation to apply.

For more information:

help(topic = "scale_x_continuous", package = "ggplot2")

transgeneNames
character. Vector indicating which assay rows denote transgenes (e.g. EGFP, TDTOMATO).

trendline
logical(1). Include trendline on plot.

tx2gene
Tx2Gene. Transcript-to-gene mappings.

unicode
logical(1). Allow Unicode characters in console output.

uploadDir
character(1). Final upload directory path.

upper
numeric(1). Upper boundary.

url
character(1). Uniform Resource Locator (URL).

value
Value to assign.

verbose
logical(1). Run the function with verbose output.

yaml
list. YAML metadata.

zeroPropagate
logical(1). Allow propagation of zeroes.

.xname
Name of object defined in x. Not intended to be used directly.

.yname
Name of object defined in y. Not intended to be used directly.

...
Additional arguments.

Value

No value.